Rename properties in the namespace to UCD
