+++ lib/Unicode/UCD/CombiningClass.hs
@@ -1,5 +1,5 @@
 -- |
--- Module      : Unicode.Properties.CombiningClass
+-- Module      : Unicode.UCD.CombiningClass
 -- Copyright   : (c) 2020 Composewell Technologies and Contributors
 -- License     : Apache-2.0
 -- Maintainer  : streamly@composewell.com
@@ -17,7 +17,7 @@
 -- You can read more about it in the Unicode standard. The latest Unicode
 -- standard can be found here: <http://www.unicode.org/versions/latest/>.
 --
-module Unicode.Properties.CombiningClass
+module Unicode.UCD.CombiningClass
     ( combiningClass
     , isCombining
     )
